pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "A Flask application for page analysis."
requires-python = ">=3.12.3"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["page_analyzer"]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
    "python-dotenv>=1.2.1",
    "validators>=0.35.0",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "beautifulsoup4>=4.14.3",
]
